chore(dashboards): display in progress/pending executions as absolute values (#13237)
Running (technically "In Progress") and pending executions were previously displayed as ratios of total executions,
which seemed to convey little value. This update switches
these KPIs to absolute values to provide a clearer, more intuitive view of current system load and renames "Running" to "In Progress" in line with other charts on the dashboards.

Co-authored-by: Georg Traar <georg@crate.io>